took out double int

+++ hello_world/6-size.c
@@ -16,10 +16,10 @@ int main(void)
 	long long int e;
 	float f;
 
-	printf("Size of char: %d byte(s)\n", (int)sizeof(c));
-	printf("Size of an int: %d byte(s)\n", (int)sizeof(i));
-	printf("Size of a long int: %d byte(s)\n", (int)sizeof(l));
-	printf("Size of a long long int: %d byte(s)\n", (int)sizeof(e));
-	printf("Size of a float: %d byte(s)\n", (int)sizeof(f));
+	printf("Size of char: %d byte(s)\n", sizeof(c));
+	printf("Size of an int: %d byte(s)\n", sizeof(i));
+	printf("Size of a long int: %d byte(s)\n", sizeof(l));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(e));
+	printf("Size of a float: %d byte(s)\n", sizeof(f));
 	return (0);
 }
